test(account-onboarding): split auto-detection suites
Split the 47-case auto-detection suite by responsibility and mock discovery at the orchestration boundary to remove unmatched MSW warnings without changing coverage.
